requirements.txt
pyspark==3.5.0
delta-spark==3.2
pandas
requests
pytest
